Use json.dump instead of json.dumps
It's arguable whether this is a better idea or not. The dumps method
potentially uses a lot more memory; it has to allocate and build the
whole JSON string before dumping it while dump writes out each bit of
JSON incrementally from (mostly) static strings. However dumps involvese
a single I/O operation, while dump involves lots.

Also used the standard json module name (which is aliased to ujson in
micropython); no point having a difference with CPython if we don't need
to.